[Windows][AArch64] Use base introspection table implementation (#442)
* Use explicit uint64_t type for HW caps

* Make MSVC cl.exe happy with introspection table macro

* [AArch64] Use base introspection table impl for Windows

* Fix missing header; remove 'MSVC-only' condition

* Fix formatting

+++ src/hwcaps.c
@@ -15,8 +15,9 @@
 #include "internal/hwcaps.h"
 
 #include <stdbool.h>
+#include <stdint.h>
 
-static bool IsSet(const uint32_t mask, const uint32_t value) {
+static bool IsSet(const uint64_t mask, const uint64_t value) {
   if (mask == 0) return false;
   return (value & mask) == mask;
 }
